refactor(e2e): restructure the recorder package for readability
Rename identifiers to say what they do, collapse the duplicated order
check, and give each file one job. No behavior change.

- createWorkload (was applyManifest), blankWithGVK (was gvkOnly),
  isStatusSettled, hasReachedTerminal, refetch (was relist, a Get not a
  LIST), attachAction.
- Unexport internals with no outside caller: classify, namedState, and the
  recording plumbing (schemaVersion, recordingPath, writeRecording,
  loadRecording, newReader, Recording.states).
- Collapse the order check: drop JourneyStep and the journeySteps adapter;
  observedOrderErr takes []journeyStep directly and is unexported.
- New doc.go (package comment) and cr.go (the unstructured-CR helpers);
  move openWatch/fireAction beside their callers in observation.go; move
  the recorder setup into recorder.go so flow.go is only the authoring DSL.

The recorder unit tests stay green.
