+++ test/e2e/recorder/order.go
@@ -7,17 +7,12 @@ import (
 	"fmt"
 	"slices"
 
-	"github.com/run-ai/karta/pkg/api/runai/v1alpha1"
+	kartav1alpha1 "github.com/run-ai/karta/pkg/api/runai/v1alpha1"
 )
 
-type JourneyStep struct {
-	State    v1alpha1.ResourceStatus
-	Optional bool
-}
-
-// ObservedOrderErr checks the observed states are a legal walk of the journey: required steps appear in
+// observedOrderErr checks the observed states are a legal walk of the journey: required steps appear in
 // order ending at want; Optional or recurring states may be absent; anything else fails.
-func ObservedOrderErr(declared []JourneyStep, observed []v1alpha1.ResourceStatus, want v1alpha1.ResourceStatus) error {
+func observedOrderErr(declared []journeyStep, observed []kartav1alpha1.ResourceStatus, want kartav1alpha1.ResourceStatus) error {
 	obs := slices.Compact(slices.Clone(observed)) // collapse consecutive repeats
 	if len(obs) == 0 {
 		return fmt.Errorf("no states observed")
@@ -25,29 +20,29 @@ func ObservedOrderErr(declared []JourneyStep, observed []v1alpha1.ResourceStatus
 	skippable := skippableSteps(declared)
 
 	// walk the observed states through the declared journey, skipping steps that may be absent
-	di := 0
+	declaredIdx := 0
 	for _, state := range obs {
-		for di < len(declared) && declared[di].State != state {
-			if !skippable[di] {
+		for declaredIdx < len(declared) && declared[declaredIdx].State != state {
+			if !skippable[declaredIdx] {
 				return fmt.Errorf("required state %q not observed before %q; journey %v, observed %v",
-					declared[di].State, state, journeyStates(declared), obs)
+					declared[declaredIdx].State, state, journeyStates(declared), obs)
 			}
-			di++
+			declaredIdx++
 		}
-		if di == len(declared) {
-			if slices.ContainsFunc(declared, func(s JourneyStep) bool { return s.State == state }) {
+		if declaredIdx == len(declared) {
+			if slices.ContainsFunc(declared, func(s journeyStep) bool { return s.State == state }) {
 				return fmt.Errorf("state %q observed out of journey %v; observed %v", state, journeyStates(declared), obs)
 			}
 			return fmt.Errorf("observed undeclared state %q; journey %v, observed %v", state, journeyStates(declared), obs)
 		}
-		di++
+		declaredIdx++
 	}
 
 	// any required steps left after the walk were never observed
-	for ; di < len(declared); di++ {
-		if !skippable[di] {
+	for ; declaredIdx < len(declared); declaredIdx++ {
+		if !skippable[declaredIdx] {
 			return fmt.Errorf("required state %q not observed; journey %v, observed %v",
-				declared[di].State, journeyStates(declared), obs)
+				declared[declaredIdx].State, journeyStates(declared), obs)
 		}
 	}
 
@@ -60,7 +55,7 @@ func ObservedOrderErr(declared []JourneyStep, observed []v1alpha1.ResourceStatus
 // skippableSteps marks each declared step that may be absent from the observed run: an Optional step, or one
 // whose state recurs earlier (compaction collapses a repeated Running from a scale, or an Initializing dip,
 // into one, so the later occurrences can never be observed on their own).
-func skippableSteps(declared []JourneyStep) []bool {
+func skippableSteps(declared []journeyStep) []bool {
 	skippable := make([]bool, len(declared))
 	for i, s := range declared {
 		skippable[i] = s.Optional
@@ -74,8 +69,8 @@ func skippableSteps(declared []JourneyStep) []bool {
 	return skippable
 }
 
-func journeyStates(steps []JourneyStep) []v1alpha1.ResourceStatus {
-	out := make([]v1alpha1.ResourceStatus, len(steps))
+func journeyStates(steps []journeyStep) []kartav1alpha1.ResourceStatus {
+	out := make([]kartav1alpha1.ResourceStatus, len(steps))
 	for i, s := range steps {
 		out[i] = s.State
 	}
